Create README.md
This project allows tourism companies to gather information from the photo-sharing social media sites like Flickr or Panoramio websites and based on the information they can draw conclusions about the tourist places and the measures they have to implement to give people a safe tour and also increase their economy. Our recommendation is to process the metadata from those photo-sharing websites and apply suitable clustering algorithms and generate the density pattern of the places that are mostly visited by the tourists, and performing regression analysis to draw up some conclusions over the site.